+++ demo/ActivityIndicatorExample.svelte
@@ -1,6 +1,6 @@
 <script>
   import Highlight from "svelte-highlight"
-  import AutoComplete from "../SimpleAutocomplete.svelte"
+  import AutoComplete from "../src/SimpleAutocomplete.svelte"
   import xml from "svelte-highlight/languages/xml"
 
   async function searchColor(keyword) {
