+++ rollup.dev.config.js
@@ -7,7 +7,7 @@ import { terser } from "rollup-plugin-terser"
 const production = !process.env.ROLLUP_WATCH
 
 export default {
-  input: "src/main.js",
+  input: "demo/main.js",
   output: {
     sourcemap: true,
     format: "iife",
