+++ src/tests/async.test.ts
@@ -1,33 +1,31 @@
 import { render, fireEvent, screen } from "@testing-library/svelte"
-import SimpleAutocomplete from "./SimpleAutocomplete.svelte"
+import SimpleAutocomplete from "../SimpleAutocomplete.svelte"
 import '@testing-library/jest-dom/extend-expect'
 import '@testing-library/jest-dom'
 
-const colors = ["White", "Red", "Yellow", "Green", "Blue", "Black", "Mät bläck", "<i>Jét Black</i>"]
+async function colors(keyword, nb_items_max) {
+    return ["White", "Red", "Yellow", "Green", "Blue", "Black", "Mät bläck", "<i>Jét Black</i>"]
+}
 
-test("items are generated but hidden", async () => {
-  const { component, container } = render(SimpleAutocomplete, {items: colors})
+test("at first, menu is empty", async () => {
+  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
 
-  expect(await screen.queryByText('White')).toBeInTheDocument()
-  expect(await screen.queryByText('White')).not.toBeVisible()
-  expect(await screen.queryByText('Red')).toBeInTheDocument()
-  expect(await screen.queryByText('Red')).not.toBeVisible()
+  expect(await screen.queryByText('White')).not.toBeInTheDocument()
+  expect(await screen.queryByText('Red')).not.toBeInTheDocument()
 })
 
-test("on focus, menu is shown with all the elements", async () => {
-  const { component, container } = render(SimpleAutocomplete, {items: colors})
+test("even with the input is focused, the menu is empty", async () => {
+  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
   const queryInput = container.querySelector("input[type='text']");
 
   await fireEvent.focus(queryInput)
 
-  expect(await screen.queryByText('White')).toBeInTheDocument()
-  expect(await screen.queryByText('White')).toBeVisible()
-  expect(await screen.queryByText('Red')).toBeInTheDocument()
-  expect(await screen.queryByText('Red')).toBeVisible()
+  expect(await screen.queryByText('White')).not.toBeInTheDocument()
+  expect(await screen.queryByText('Red')).not.toBeInTheDocument()
 })
 
 test("when something is queried, only the matching items are shown", async () => {
-  const { component, container } = render(SimpleAutocomplete, {items: colors})
+  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
   const queryInput = container.querySelector("input[type='text']");
 
   await fireEvent.input(queryInput, { target: { value: "white" } })
@@ -38,7 +36,7 @@ test("when something is queried, only the matching items are shown", async () =>
 })
 
 test("when nothing matches the query, no item is show", async () => {
-  const { component, container } = render(SimpleAutocomplete, {items: colors})
+  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
   const queryInput = container.querySelector("input[type='text']");
 
   await fireEvent.input(queryInput, { target: { value: "not-a-color" } })
@@ -48,7 +46,7 @@ test("when nothing matches the query, no item is show", async () => {
 })
 
 test("when something is queried, the query is highlighted", async () => {
-  const { component, container } = render(SimpleAutocomplete, {items: colors})
+  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
   const queryInput = container.querySelector("input[type='text']");
   const list = container.querySelector("autocomplete-list");
