+++ src/tests/sync.test.ts
@@ -1,31 +1,33 @@
 import { render, fireEvent, screen } from "@testing-library/svelte"
-import SimpleAutocomplete from "./SimpleAutocomplete.svelte"
+import SimpleAutocomplete from "../SimpleAutocomplete.svelte"
 import '@testing-library/jest-dom/extend-expect'
 import '@testing-library/jest-dom'
 
-async function colors(keyword, nb_items_max) {
-    return ["White", "Red", "Yellow", "Green", "Blue", "Black", "Mät bläck", "<i>Jét Black</i>"]
-}
+const colors = ["White", "Red", "Yellow", "Green", "Blue", "Black", "Mät bläck", "<i>Jét Black</i>"]
 
-test("at first, menu is empty", async () => {
-  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
+test("items are generated but hidden", async () => {
+  const { component, container } = render(SimpleAutocomplete, {items: colors})
 
-  expect(await screen.queryByText('White')).not.toBeInTheDocument()
-  expect(await screen.queryByText('Red')).not.toBeInTheDocument()
+  expect(await screen.queryByText('White')).toBeInTheDocument()
+  expect(await screen.queryByText('White')).not.toBeVisible()
+  expect(await screen.queryByText('Red')).toBeInTheDocument()
+  expect(await screen.queryByText('Red')).not.toBeVisible()
 })
 
-test("even with the input is focused, the menu is empty", async () => {
-  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
+test("on focus, menu is shown with all the elements", async () => {
+  const { component, container } = render(SimpleAutocomplete, {items: colors})
   const queryInput = container.querySelector("input[type='text']");
 
   await fireEvent.focus(queryInput)
 
-  expect(await screen.queryByText('White')).not.toBeInTheDocument()
-  expect(await screen.queryByText('Red')).not.toBeInTheDocument()
+  expect(await screen.queryByText('White')).toBeInTheDocument()
+  expect(await screen.queryByText('White')).toBeVisible()
+  expect(await screen.queryByText('Red')).toBeInTheDocument()
+  expect(await screen.queryByText('Red')).toBeVisible()
 })
 
 test("when something is queried, only the matching items are shown", async () => {
-  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
+  const { component, container } = render(SimpleAutocomplete, {items: colors})
   const queryInput = container.querySelector("input[type='text']");
 
   await fireEvent.input(queryInput, { target: { value: "white" } })
@@ -36,7 +38,7 @@ test("when something is queried, only the matching items are shown", async () =>
 })
 
 test("when nothing matches the query, no item is show", async () => {
-  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
+  const { component, container } = render(SimpleAutocomplete, {items: colors})
   const queryInput = container.querySelector("input[type='text']");
 
   await fireEvent.input(queryInput, { target: { value: "not-a-color" } })
@@ -46,7 +48,7 @@ test("when nothing matches the query, no item is show", async () => {
 })
 
 test("when something is queried, the query is highlighted", async () => {
-  const { component, container } = render(SimpleAutocomplete, {searchFunction: colors})
+  const { component, container } = render(SimpleAutocomplete, {items: colors})
   const queryInput = container.querySelector("input[type='text']");
   const list = container.querySelector("autocomplete-list");
